Introduction to Test Driven Development (TDD)/labs/04_test_coverage/requirements.txt
# Dependencies for this lab
Flask
Flask-SQLAlchemy
nose
pinocchio
coverage
pylint
